ref: extract utils methods in sub-modules
