+++ utils/__init__.py
@@ -1,9 +1,12 @@
 # flake8: noqa
-from .utils import (
+
+from .embeddings import (
+    ClipImageEmbedding,
+    NoEmbedderForDocumentTypeException,
     embedd_image,
     embedd_text,
-    local_image_to_data_url,
-    pil_to_bytes,
-    singleton,
-    split_in_chunks,
+    get_embedder,
 )
+from .images import b64encode_image, local_image_to_data_url, open_image, pil_to_bytes
+from .lists import split_in_chunks
+from .utils import singleton
